Rename Project to Blaze
We have to rename our project, because using FHIR in product names isn't allowed
without permission. Because we don't like to apply for permission at this stage
of the project, we rename it to Blaze.

From: https://www.hl7.org/fhir/license.html

Use of Health Level Seven International trademarks in URL domains or to brand
your product or service without the express written consent of Health Level
Seven International is strictly prohibited.