Output lastUpdated in Meta
With this issue, I replaced the lazy resource with a combination of a
resource-handle and a pull function. Doing so has the advantage that we
no longer do a blocking fetch of the resource contents inside the lazy
resource. Instead we can now pull the full resource if we need to. This
pull function is async. So we can decide wether to pull blocking or not.
In CQL we pull blocking. In the REST API we pull async.

Closes: #190