Improve Transaction Performance
We relied on the resource cache to transfer resources from the submit
side of the transaction handling to the resource indexing side. However
a cache isn't well suited to hold entries that aren't already accessed.
So the hit ratio wasn't near 100 % as it should for this use case.

To improve on this situation, we now maintain a separate resource cache
in the database node.

Also did many more improvements and updates.