Fix Parallel Transaction Creation Test
The test was failing in CI. That's because it's not guaranteed that
parallel calls to `d/transact` will result in ordered return of db's
with increasing t's.

I rewrote the test using `d/as-of` for obtaining databases with all t's
since start. Since we create one patient in every transaction, the
number of patients has to be equal to the t of the database.