Fix Indexing of lastUpdated for Deleted Resources
When deleting a resource, a resource content with type and id only was
stored and indexed. However, because the lastUpdated timestamp is not
part of the resource content, it was associated to the resource content
before indexing. So the lastUpdate timestamp was indexed even for the
deleted version of a resource, so that a search for _lastUpdated found
that version and returned it as match in results.

This patch completely omits storing and indexing of deleted resource
contents and uses the hash consisting of all zero bytes for the
resource handle. Pulling such a deleted resource handle will return a
resource content consisting of type and id only immediately.