Fix Memory Leak in Node
The node had a transaction resource cache that stores resources of a
transaction by t. The node puts the resources after the transaction was
submitted successfully. However sometimes the node gets back the tx-data
from the transaction log before it puts the resources in the cache.

I removed this cache completely and send the resources as local payload
through the local transaction log. The Kafka transaction log on the
other hand doesn't transfer the resources.