Allow Setting the Database Sync Timeout
All reading FHIR interactions have to acquire the last database state
known at the time the request arrived in order to ensure consistency.
That database state might not be ready immediately because indexing
might be still undergoing. In such a situation, the request has to wait
for the database state becoming available. If the database state won't
be available before the timeout expires, a 503 Service Unavailable
response will be returned. Please increase this timeout if you
experience such 503 responses, and you are not able to improve indexing
performance or lower transaction load.