Improve Test Coverage Execution
Because we can execute the test coverage only sequentially, it is more
efficient to do this in a single job rather than on multiple.

Also improved cassandra module test coverage in order to not lower the
overall test coverage because the cassandra module was missing from the
current job listing.