Intern Uri, Canonical, Code, Coding and CodeableConcept
In FHIR, some values repeat in many resources. First uri's are used as
system's in Codings and Identifiers. Second canonical values are used to
refer to other resources. Third codes are often the same for a large
number of resources. Last, Coding and CodeableConcept contain system and
code values and are also often the same in it's entirety.

By interning instances of those values, only one instance will resist in
memory as opposed to many. Because all values are immutable, there is no
risk in sharing instances. The memory savings are about 50 % for Synthea
BMI Observations.