Fix Server Errors Because of Unencoded Error Outputs
This problem was introduced in v0.17.1. Versions of Blaze before v0.17.1
are not effected.

HTTP Error Outputs with status codes 404 and 405 were not encoded to
either JSON or XML, leading to an error in Jersey, leading to a 500.

This commit also introduces redirecting requests including a leading
slash to ones without one.