Fix Content Negotiation
This fixes a regression introduced by #703 where we don't return JSON
by default if an Accept header is present. This is relevant in
situations where browsers start with text/html but also will accept
application/xml with lower priority. Instead of returning a 406 in this
cases, XML is delivered now.

If one prefers JSON in browsers the _format query param has to be used.