Fix Git SHA Problems in GitHub Actions
The variable github.sha points now to the merge commit of a PR but the
Docker metadata action uses the SHA of the commit to merge.

Now in addition to pushing the Docker image into the registries, I build
local Docker image and upload it to the artifact storage to be able to
use it at the integration tests.