Fix Storage of Bundles with References
Before, references inside Bundles were tried to resolve during
transactions. That required resources to exists that were referenced
inside Bundles. But Bundles should be stored "as-is", opaque without
resolving any references in them.

This fix, changes the function blaze.fhir.spec.type/references to not
look inside Bundle entries. The integration test
"bundle-with-references-test" was added to try to issue a transaction
that stores a transaction bundle with a reference in it.