Fix CQL Function Argument Hiding
I added CQL functions in #501. Function arguments are added to the
scope. However in queries, I had an optimization that overrode the scope
with a single value if the scope was empty. The problem was that I did
not pay attention to the function argument scope. To keeps things
simple, I just removed the optimization.