Prepare Transaction Log API to Transport Resources
Before that change the transaction log API handled only resource hashes
instead of whole resources. The resources itself had to be handled
outside of the transaction log.

Now the API of the transaction log takes resources instead of hashes.
Internally the local and the Kafka implementation still use the
resource store in order to transport the resources out-of-band so that
the transaction log storage itself doesn't have to contain the whole
resources.

This change enables transaction log implementations that transport the
resources in-band and so don't need a resource store. At least not for
the resources that are not already indexed.

This change is fully compatible with the current storage model and so is
a pure refactoring.