Fix _lastUpdated Search Returning a Resource more than Once
Searching resources by _lastUpdated can return a resource more than
once.

The problem with the meta.lastUpdated property is that it doesn't
contribute to the content hash of a resource because it's not part of
the resource content. This results in multiple entries into the
SearchParamValueResource index having the same content hash. If that
content hash is equal to the current one, the resource is returned
multiple times.

The fix moves the _lastUpdated search clause away from the first
position because the first clause uses the SearchParamValueResource
index.

Closes: #882