Allow Writing Large Binary Resources in XMl Format
The Woodstox XML Parser has a default limit of 524288 bytes for XML
attribute values in order to prevent DOS attacks. However for FHIR
Binary resources that limit is just to low. I decided against a
configurable limit, because the possibility to do so is implementation
specific to Woodstox and I can't guarantee that I can implement it
using other libraries. So I disabled the limit completely.