Improve Performance Evaluating Measures without Stratifier
Evaluating a Measure with reportType of population and no stratifier
doesn't need collect the handles. This change determines that situation
and just calculates the population count instead of collecting all
handles.

This changes enables Blaze to determine the count of large populations
were previously lots of memory was used for no reason.

The speedup documented in docs/performance/cql.md is not from this
change. I improved the cache sizes and GC in order to achieve the
performance gains.

I added the observation-category-laboratory query to show that we can
now count 170 M observations without using heap memory.