Improve Date Search Performance
Used indexes for equal, less than and starts after. That operators all
have a predicate on the lower bound of the resource value that comes
first in the index. The other operators can't use the existing index,
because they either look at the upper bound or on both bounds.

We have to add indexes in order to speed up the other operators. The
issue is: #976.