Fix Extended Birth Date in CQL
For age calculation in CQL, the property birthDate.value is used. That
.value part means that we have to get the value of the FHIR date type.
We never did this and so got ExtendedDate into the CQL evaluation which
doesn't work.

Also overhauled the System.Date and System.DateTime types. Decided to
implement them in Java. Improved the parsing performance by about 4
times. For Date/DateTime up to Date precision.

Closes: #973