Disable Resource Handle Cache if Max-Size Zero
Before, if one sets the max-size of the resource handle cache to zero,
a cache was still created which results in a certain overhead on each
lookup. Now this overhead is eliminated because no cache is created and
the lookup is performed directly into the RocksDB index.

Also moved the implementation of the Key class into a Java source file.