Fix Non-Stable Paging
I was actually not using an as-of database for requests with a __t
query param. So although the database was synced on t, as sync returns
the newest database with at least t, we were operating on a to new
database if writes happened in between. A simple as-of call to the
database will do the trick to get a database with exactly t.