Remove Shell in Docker Entrypoint
We used a shell in the docker entrypoint in order to be able to specify
the absolute path to libjemalloc.so.2. However that isn't necessary as
documented in https://man7.org/linux/man-pages/man8/ld.so.8.html. So we
switch back to a direct java -jar call in the entrypoint.

Note that this will not change the runtime behaviour of Blaze because
in the shell, we used exec to break out of the shell and have the java
process running as the only process in the container.

Closes: #1034