Update Netty to Resolve CVE-2023-34462
This vulnerability isn't relevant for Blaze because we use Netty only
for client connections to Cassandra. The risk that Cassandra would work
as an attack vector for DOS attacks agains Blaze is considered very low.

So this security fix doesn't demand an immediate release of a security
fix version of Blaze.