Improve Dependency Graph Breaking Up test-util
Modules like anomaly needed a dependency to fhir-structure in order to
facilitate prepping of code that was only needed from test-util. By
moving the dependency on fhir-structure into a separate fhir-test-util
module, the test-util module is free from the dependency on
fhir-structure and modules that don't depend on fhir-structure don't
need to do that only for testing purposes.