Move Test Coverage into its own Workflow
Calculating the test coverage is independent from all other parts of the
build workflow. It also takes the most of the time and so it would be
good to have it run parallel to all other jobs. Putting it into its own
Workflow also makes sense because we like to have it run only on PR's
towards develop.