Simplify with-open-coll Macro and coll/eduction
Both the with-open-coll macro and the coll/eduction function implemented
Seqable, the with-open-coll macro also Counted. Both interfaces were
only used in tests or could be replaced by IReduceInit handling
functions in production. Not implementing both interfaces saves a lot of
test efforts, especially for the with-open-coll macro because all the
forms have to be tested for each occurrence of the macro.

The only disadvantage is that the collections now only implement
IReduceInit or both IReduceInit and Counted in case of coll/eduction,
which means that they can't be used in places were sequences are needed.
However this is also a performance advantage because the sequence
abstraction implemented in the with-open-coll macro always consumed the
full collection. That means that a (take 2 coll) would consume the
whole coll were a (into [] (take 2) coll) would only consume the first
two elements. Now that (take 2 coll) will no longer work, the more
efficient variant is enforced.