Allow Handling of Large CQL Libraries
Currently certain large CQL libraries can't be compiled because the
translated ELM representation in JSON contains objects nested at a
depth larger than 1000. There is a max nesting depth of 1000 in the
JsonFactory used by the Jackson ObjectMapper. In this issue we like to
increase that max nesting dept to 5000. If there are still reports of
even larger queries not compilable, we would make the max depth a
config option. But for now we don't like to expose an additional config
option.

Closes: #1268