Refactor d/patient-everything Returning the Patient
We need the patient anyway in patient everything, but we added it later
after taking the max 10.0000 resources. Now in #1297 we will expand the
linked patients, so we will need to return the patients from
d/patient-everything because there could be more than the initial
patient.

I use the cat transducer and an extra eduction to concatenate the
patient-handle with the other resource handles.