Refactor $evaluate-measure Execution
We now use completable futures instead of reducers to evaluate CQL
expressions parallel.

This commit also brings a performance boost because the new parallel
execution architecture doesn't keep the order of evaluation results
in sync with the initial patient order. Because we don't need to keep
the order of the evaluation results, the worker threads can be utilized
better and so the performance increases. For 32 core systems the
performance increase is about 50%. For systems with letter cores, the
increase is less.

Closes #1032 because we no longer use reducers at all.