Remove Custom API for Default Column Family
RocksDB has a custom API for the default column family from the time
before column family support. We used that custom api also in the Key
Value Store abstraction, because we didn't know that it's the same as
accessing the default column family.

This change simplifies the Key Value Store API.