Improve other-clauses-filter Function
Before one filter xform was created that looped over the clauses, now
a function composition of one filter per clause is created. The new
xform will still short cut after the first non-match and has the
advantage that the clauses do not have to be looped over in every filter
step.