Remove Database T from Self Links
Self links should point to the location were a new FHIR search can be
started and not to the immutable page of the current search result.

This also means that only the first page will contain a self link.
Furthermore the Search Type and Search System endpoint will no longer
need to obtain a database with a certain t. So databases with a certain
t will only needed to be obtained by __page endpoints and vread. This
makes the surface area of getting access to older database states
smaller.