Suppress Log Output at Code Coverage Runs
While running normal unit tests, the log output is captured by Kaocha.
However for code coverage runs via Cloverage that was not the case.
Having such enormous log output is bad for GitHub actions and
potentially slower. Using the Cloverage plugin of Kaocha, we are now
able to capture the log output also for coverage runs.