Cleanup Codes Evaluation in Retrieve Expression
In case the codes of a Retrieve expression were obtained from a Concept,
the source-property expression that get the codes of a Concept was
preserved until the clauses had to be calculated. Now the
source-property expression will evaluate the codes path on the static
Concept during compilation as other expressions also do. After the codes
expression is compiled, it's checked for holding a list of codes. So
only static codes expressions are supported now explicitly and an
appropriate exception is thrown otherwise.

Doing so ensures that the codes argument used in external-data is
actually a list of codes and so we also can calculate the query clauses
in a simple way.

That allowed to remove the record for the source-property expression. So
the codes is simpler and safer as before.