Optimize Thread Usage while Loading Resources
Before this change, the KV Resource Store used it's thread pool for both
loading and parsing/conforming the resources. However loading is I/O but
parsing/conforming is CPU bound. So it is better to use the KV Resource
Store Executor only for the loading, the I/O part.

With this change the KV Resource Store uses do-async to change to a
common ForkJoinPool thread. This change has also the advantage that
functions that are applied after the returned future do not have to
change the thread pool because the futures are already on the common
ForkJoinPool.

The performance improves by a little bit because the threads are now
directly usable to lead the next resource instead having to parse/
conform the current one.

Removed all pool switching async functions. Tested the thread pool used.