Reintroduce Configurable Thread Pool for Measure Evaluation
Because accessing the contents of a resource isn't async in measure
evaluation, the thread pool can't be CPU bound only. In case a resource
has to be loaded from disk the measure evaluation thread will be
blocked and so can't be used for computation anymore. Also modern SSDs
need a certain amount of I/O requests in flight in order to saturate the
drive. If the number of available processors is lower than the number
of parallel I/O requests needed, the storage can't be even used fully
with a CPU bound pool.

So the old env var FHIR_OPERATION_EVALUATE_MEASURE_THREADS is available
again.