Refactor How Purging Reflects in Resource Handles
Before this commit, resource handles had the t at which it was purged
as field called purged-at. That field was used later to decide whether
a resource handle should be part of a history at a particular t.

After this commit, the resource-handle! function will take an additional
argument called base-t and will only return resource handles not purged
at this base-t. Doing so, there will be never a purged resource handle
and so doesn't need to be checked at all. So the field purged-at is
gone.