Improve Anomaly Detection Performance
Profiling the new parser has shown that about 15% is spend detecting
whether a value is an anomaly or not. Especially records like Coding or
Quantity have poor performance on keyword lookup. The new version first
tests that the anomaly is actually a real Clojure map (APersistentMap)
which a record isn't and than even uses .containsKey to save recurring
type checks.

Closes: #2506