Remove Choice Type Cache from Module fhir-structure
Remove a cache of the rather simple calculation

```clojure
(keyword (str (name key) (su/capital (name type))))
```

I've found nothing in the git history that proves the effectiveness of
this cache. Also after #2488 that cache was only used for XML processing
which is not as widely used as JSON.

Without this cache, we can remove the caffeine dependency from
fhir-structure.