Fix Consent Resource policyRule Property
The problem was that the property `policyRule` comes after the property
`policy` in the list of element definitions. Both have the same prefix
and `policy` is a BackboneElement. The function
separate-element-definitions* used string prefix on path as termination
condition, so `policyRule` was counted as member of `policy`. I just
added a `.` to make the prefix check robust.

Closes: #2700