Fix Key Shape of Terminology Service Graph Cache
The key can't be a string concatenated from id and meta.versionId (t)
because the chars allowed in t are also valid in the id. So for example
the id `A12` and t `3` would yield the same key as the id `A1` and the t
`23`.

We could use a separator like `|` but just using a vector is simpler.