Improve String Concatenation Performance
The function clojure.core/str that is usually used for string
concatenation has a performance problem which originates in missing
arities. The str function has only the arities 0, 1, and variadic. That
means that concatenating two strings will use the variadic arity which
will put the second argument into a list before the variadic version is
called. The implementation uses a StringBuilder that is initialized with
the first string while all other strings are appended. Doing so will
result in a StringBuilder eventually having to resize it's internal byte
array. On the other hand, if the str function had multiple arities,
individual known strings can be concatenated.