AOT Compile all Namespaces
We like to compile all namespaces ahead of time (AOT) because we can
ensure direct linking and improve the startup time of Blaze.

The compilation step was always expected to compile every namespace,
however it didn't. The reason was that all the integrant component
namespaces are not required in blaze.core. We do that because we don't
like to namespaces of disabled features. Instead the namespaces of
components actually in use are loaded via integrant itself.

Because the compiler doesn't see the namespaces of components, we have
to give it all that namespaces.

The key :src-dirs is actually not needed, because compile-clj already
uses all source dirs from basis.