Test Queries with Multiple Codes and Patients
Before we had performance and functional tests for queries with a single
code like a laboratory observation with a certain LOINC code over
multiple patients.

This commit adds one functional test and a performance test for queries
with multiple codes in addition to multiple patients.

For the functional download tests with query, I added an additional
check of resource uniqueness.