Fix Stateful Transducers
Some stateful transducers created their state at transducer creation
time which is wrong. The state has to be created at the time the
reduce function is applied.

In case the collection is reduced only one that makes no difference.
However if it is reduced more than one, the state is reused.