Fix Next Link Detection
In R5 the link relation in a bundle is a code instead of a string like
in R4. Because a string without extensions is just a string itself,
calling type/value isn't needed. But in R5 we need to call type/value
because a code isn't a string.

Also fixed reflection warnings in test.