Extract Bundle Link Creation into a Function
The function will create bundle links depending of the FHIR version in
use. This is the first function doing so. The function is build as
component which takes the FHIR version at initialization and returns an
appropriate function. Although the implementation for FHIR
v6.0.0-ballot3 is already given, and could be even used by setting the
env var FHIR_VERSION accordingly, this is not supported until full R6
availability is documented.