Improve JSON Parsing Performance
Use singly-linked lists instead of vectors for the locator because
adding a path segment in front if cheaper than appending it to the end
of a vector. Reversing the list is no problem, because we don't optimize
for the error case.

Use mutable ArrayLists instead of transient vectors/maps because
creating transient versions of vectors/maps is just expensive.

Close the TreeTraversingParser because I just forgot that.