Refactor Resource Handle
Resource handle is now a Java class with keyword lookup sites. They
allow for efficient keyword lookup of the data of resource handle. Gone
are all access functions in the resource-handle namespace.

Removed the FhirType protocol from resource handle and CQL resource.
Fhir type information is accessed by the :fhir/type keyword.

Keys of the resource store now contain the fhir type of the resource as
keyword like :fhir/Patient instead of the string "Patient". This ensures
that the type part of the keys is always interned.