Refactor Resource Cache
Before this change, the resource cache implemented the resource store
protocol. So the resource cache was a transparent layer on top of the
resource store. Normally this is a good idea but in #3086 we need
additional functionality in the resource cache.

This refactoring gives the resource cache it's own protocol and set of
access functions. The database node needs now both the resource cache
and the resource store, because the resource cache has no longer the
write functionality.

Also in case of a resource cache size of zero, a real no-op cache is
returned.